% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromosome1.R
\docType{data}
\name{Chromosome1}
\alias{Chromosome1}
\title{Chromosome1}
\format{
Large vcfR object containing 3 data.frames for metadata, fix data, and genotype data
}
\usage{
Chromosome1
}
\description{
A dataset containing a subset of WGS vcf. The data is from a population of Akagera zebras that
that contain striped and spotted zebras within the population.
}
\keyword{datasets}
