% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HpFst_final.R
\name{HoldFst}
\alias{HoldFst}
\title{Heterozygosity and Fixation Index Scores}
\usage{
HoldFst(file, P, Q, bin = 5e+05, plot = TRUE)
}
\arguments{
\item{file}{file path to vcf file containing both the normal and variant population}

\item{P}{The name of the normal population (ex. 'Normal')}

\item{Q}{The name of the population with the phenotypic difference (ex. 'Variant')}

\item{bin}{Size of windows to calculate Heterozygosity and Fst scores over (default=500kb)}

\item{plot}{Logical. Should the function return manhattan plots (T/F)}
}
\description{
Calculate Heterozygosity scores and Fixation Index for low-coverage whole genome sequencing (Description)
}
\details{
Calculate and plot scores from low-coverage sequencing from a vcf input file (Details)
}
\examples{
#HoldFst('input.vcf','Normal','Variant',500000,plot=T)
}
\author{
Ashlyn Anderson
}
