% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopWindows.R
\name{topwindows}
\alias{topwindows}
\title{Find top windows of the genome associated with a difference between the normal and variant population}
\usage{
topwindows(object, sig = "all")
}
\arguments{
\item{object}{The object containing the results from HoldFst}

\item{sig}{Which scores should be used to find the most significant windows ('Fst'=Fixation Index, 'Phet'=heterozygosity of P population,'Qhet'=heterozygosyity of Q population,'Diff'=difference in heterozygosity,'All'=all scores(default))}
}
\description{
Top Windows for Heterozygosity and Fixation  (Description)
}
\details{
Input results from running HoldFst to get the top most significant windows (Details)
}
\examples{
#topwindows(results,sig='Qhet')

}
\author{
Ashlyn Anderson
}
